Not all the monitor code got the memo about nodeid being a bigint. (#729)
* Not all the monitor code got the memo about nodeid being a bigint.

Some of the code was still expecting nodeid to fit in 32 bits, where in the
SQL schema it's been changed to be a bigint/bigserial. On 64 bits systems,
it seems like we've been saved by alignment properties.

* Assorted bug fixes on the monitor client.

* Fix client-side code to also use 64 bits integers for the nodeId.

Except in the state file, which would require more coding that is felt
necessary at the moment.

* Fix int32 nodeId in the health check code.

* Fix string formats for the monitor extension code, use %ld.

It seems to be what Postgres does and expects.

* Run make installcheck on i386, automate that, and make it pass.

* Assorted fixes.

* Review indentation when using constants PRId64/NODE_FORMAT.

* Add missed drop statement

* Use int64 apis in a bit more places

+++ src/monitor/group_state_machine.h
@@ -22,7 +22,7 @@
  */
 typedef struct AutoFailoverNodeState
 {
-	int32 nodeId;
+	int64 nodeId;
 	int32 groupId;
 	ReplicationState replicationState;
 	XLogRecPtr reportedLSN;
