+++ src/monitor/pgautofailover.sql
@@ -236,14 +236,14 @@ CREATE FUNCTION pgautofailover.register_node
     IN dbname               name,
     IN node_name            text default '',
     IN sysidentifier        bigint default 0,
-    IN desired_node_id      int default -1,
+    IN desired_node_id      bigint default -1,
     IN desired_group_id     int default -1,
     IN initial_group_role   pgautofailover.replication_state default 'init',
     IN node_kind            text default 'standalone',
     IN candidate_priority 	int default 100,
     IN replication_quorum	bool default true,
     IN node_cluster         text default 'default',
-   OUT assigned_node_id     int,
+   OUT assigned_node_id     bigint,
    OUT assigned_group_id    int,
    OUT assigned_group_state pgautofailover.replication_state,
    OUT assigned_candidate_priority 	int,
@@ -254,7 +254,7 @@ RETURNS record LANGUAGE C STRICT SECURITY DEFINER
 AS 'MODULE_PATHNAME', $$register_node$$;
 
 grant execute on function
-      pgautofailover.register_node(text,text,int,name,text,bigint,int,int,
+      pgautofailover.register_node(text,text,int,name,text,bigint,bigint,int,
                                    pgautofailover.replication_state,text,
                                    int,bool,text)
    to autoctl_node;
@@ -263,13 +263,13 @@ grant execute on function
 CREATE FUNCTION pgautofailover.node_active
  (
     IN formation_id           		text,
-    IN node_id        		        int,
+    IN node_id        		        bigint,
     IN group_id       		        int,
     IN current_group_role     		pgautofailover.replication_state default 'init',
     IN current_pg_is_running  		bool default true,
     IN current_lsn			  		pg_lsn default '0/0',
     IN current_rep_state      		text default '',
-   OUT assigned_node_id       		int,
+   OUT assigned_node_id       		bigint,
    OUT assigned_group_id      		int,
    OUT assigned_group_state   		pgautofailover.replication_state,
    OUT assigned_candidate_priority 	int,
@@ -279,7 +279,7 @@ RETURNS record LANGUAGE C STRICT SECURITY DEFINER
 AS 'MODULE_PATHNAME', $$node_active$$;
 
 grant execute on function
-      pgautofailover.node_active(text,int,int,
+      pgautofailover.node_active(text,bigint,int,
                           pgautofailover.replication_state,bool,pg_lsn,text)
    to autoctl_node;
 
@@ -287,7 +287,7 @@ CREATE FUNCTION pgautofailover.get_nodes
  (
     IN formation_id     text default 'default',
     IN group_id         int default NULL,
-   OUT node_id          int,
+   OUT node_id          bigint,
    OUT node_name        text,
    OUT node_host        text,
    OUT node_port        int,
@@ -307,7 +307,7 @@ CREATE FUNCTION pgautofailover.get_primary
  (
     IN formation_id      text default 'default',
     IN group_id          int default 0,
-   OUT primary_node_id   int,
+   OUT primary_node_id   bigint,
    OUT primary_name      text,
    OUT primary_host      text,
    OUT primary_port      int
@@ -323,8 +323,8 @@ grant execute on function pgautofailover.get_primary(text,int)
 
 CREATE FUNCTION pgautofailover.get_other_nodes
  (
-    IN nodeid           int,
-   OUT node_id          int,
+    IN nodeid           bigint,
+   OUT node_id          bigint,
    OUT node_name        text,
    OUT node_host        text,
    OUT node_port        int,
@@ -334,17 +334,17 @@ CREATE FUNCTION pgautofailover.get_other_nodes
 RETURNS SETOF record LANGUAGE C STRICT
 AS 'MODULE_PATHNAME', $$get_other_nodes$$;
 
-comment on function pgautofailover.get_other_nodes(int)
+comment on function pgautofailover.get_other_nodes(bigint)
         is 'get the other nodes in a group';
 
-grant execute on function pgautofailover.get_other_nodes(int)
+grant execute on function pgautofailover.get_other_nodes(bigint)
    to autoctl_node;
 
 CREATE FUNCTION pgautofailover.get_other_nodes
  (
-    IN nodeid           int,
+    IN nodeid           bigint,
     IN current_state    pgautofailover.replication_state,
-   OUT node_id          int,
+   OUT node_id          bigint,
    OUT node_name        text,
    OUT node_host        text,
    OUT node_port        int,
@@ -355,11 +355,11 @@ RETURNS SETOF record LANGUAGE C STRICT
 AS 'MODULE_PATHNAME', $$get_other_nodes$$;
 
 comment on function pgautofailover.get_other_nodes
-                    (int,pgautofailover.replication_state)
+                    (bigint,pgautofailover.replication_state)
         is 'get the other nodes in a group, filtering on current_state';
 
 grant execute on function pgautofailover.get_other_nodes
-                          (int,pgautofailover.replication_state)
+                          (bigint,pgautofailover.replication_state)
    to autoctl_node;
 
 CREATE FUNCTION pgautofailover.get_coordinator
@@ -411,15 +411,15 @@ grant execute on function pgautofailover.get_most_advanced_standby(text,int)
 
 CREATE FUNCTION pgautofailover.remove_node
  (
-   node_id int
+   node_id bigint
  )
 RETURNS bool LANGUAGE C STRICT SECURITY DEFINER
 AS 'MODULE_PATHNAME', $$remove_node_by_nodeid$$;
 
-comment on function pgautofailover.remove_node(int)
+comment on function pgautofailover.remove_node(bigint)
         is 'remove a node from the monitor';
 
-grant execute on function pgautofailover.remove_node(int)
+grant execute on function pgautofailover.remove_node(bigint)
    to autoctl_node;
 
 CREATE FUNCTION pgautofailover.remove_node
@@ -464,24 +464,24 @@ comment on function pgautofailover.perform_promotion(text,text)
 grant execute on function pgautofailover.perform_promotion(text,text)
    to autoctl_node;
 
-CREATE FUNCTION pgautofailover.start_maintenance(node_id int)
+CREATE FUNCTION pgautofailover.start_maintenance(node_id bigint)
 RETURNS bool LANGUAGE C STRICT SECURITY DEFINER
 AS 'MODULE_PATHNAME', $$start_maintenance$$;
 
-comment on function pgautofailover.start_maintenance(int)
+comment on function pgautofailover.start_maintenance(bigint)
         is 'set a node in maintenance state';
 
-grant execute on function pgautofailover.start_maintenance(int)
+grant execute on function pgautofailover.start_maintenance(bigint)
    to autoctl_node;
 
-CREATE FUNCTION pgautofailover.stop_maintenance(node_id int)
+CREATE FUNCTION pgautofailover.stop_maintenance(node_id bigint)
 RETURNS bool LANGUAGE C STRICT SECURITY DEFINER
 AS 'MODULE_PATHNAME', $$stop_maintenance$$;
 
-comment on function pgautofailover.stop_maintenance(int)
+comment on function pgautofailover.stop_maintenance(bigint)
         is 'set a node out of maintenance state';
 
-grant execute on function pgautofailover.stop_maintenance(int)
+grant execute on function pgautofailover.stop_maintenance(bigint)
    to autoctl_node;
 
 CREATE FUNCTION pgautofailover.last_events
@@ -690,7 +690,7 @@ CREATE OR REPLACE FUNCTION pgautofailover.update_secondary_check()
   LANGUAGE 'plpgsql'
 AS $$
 declare
-  nodeid        integer := null;
+  nodeid        bigint := null;
   reportedstate pgautofailover.replication_state := null;
 begin
 	-- when secondary changes from true to false, check all nodes remaining are primary
